final-project/News-Scrap-main/requirements.txt
fastapi[standard]
requests
sqlalchemy
pydantic
python-dotenv
mysql-connector-python